+++ lib/Migration/InstallBot.php
@@ -26,23 +26,16 @@ declare(strict_types=1);
 
 namespace OCA\CallSummaryBot\Migration;
 
-use OCA\CallSummaryBot\Model\Bot;
+use OCA\CallSummaryBot\Service\BotService;
 use OCA\Talk\Events\BotInstallEvent;
-use OCP\EventDispatcher\IEventDispatcher;
-use OCP\IConfig;
 use OCP\IURLGenerator;
-use OCP\L10N\IFactory;
 use OCP\Migration\IOutput;
 use OCP\Migration\IRepairStep;
-use OCP\Security\ISecureRandom;
 
 class InstallBot implements IRepairStep {
 	public function __construct(
-		protected IConfig $config,
-		protected IEventDispatcher $dispatcher,
-		protected ISecureRandom $random,
 		protected IURLGenerator $url,
-		protected IFactory $l10nFactory,
+		protected BotService $service,
 	) {
 	}
 
@@ -57,45 +50,6 @@ class InstallBot implements IRepairStep {
 		}
 
 		$backend = $this->url->getAbsoluteURL('');
-		$id = sha1($backend);
-
-		$secretData = $this->config->getAppValue('call_summary_bot', 'secret_' . $id);
-		if ($secretData) {
-			$secretArray = json_decode($secretData, true, 512, JSON_THROW_ON_ERROR);
-			$secret = $secretArray['secret'] ?? $this->random->generate(64, ISecureRandom::CHAR_HUMAN_READABLE);
-		} else {
-			$secret = $this->random->generate(64, ISecureRandom::CHAR_HUMAN_READABLE);
-		}
-
-		foreach (Bot::SUPPORTED_LANGUAGES as $lang) {
-			$this->installLanguage($secret, $lang);
-		}
-
-		$this->config->setAppValue('call_summary_bot', 'secret_' . $id, json_encode([
-			'id' => $id,
-			'secret' => $secret,
-			'backend' => $backend,
-		], JSON_THROW_ON_ERROR));
-	}
-
-	protected function installLanguage(string $secret, string $lang): void {
-		$libL10n = $this->l10nFactory->get('lib', $lang);
-		$langName = $libL10n->t('__language_name__');
-		if ($langName === '__language_name__') {
-			$langName = $lang === 'en' ? 'British English' : $lang;
-		}
-
-		$l = $this->l10nFactory->get('call_summary_bot', $lang);
-
-		$event = new BotInstallEvent(
-			$l->t('Call summary'),
-			$secret . str_replace('_', '', $lang),
-			$this->url->linkToOCSRouteAbsolute('call_summary_bot.Bot.receiveWebhook', ['lang' => $lang]),
-			$l->t('Call summary (%s)', $langName) . ' - ' . $l->t('The call summary bot posts an overview message after the call listing all participants and outlining tasks'),
-		);
-		try {
-			$this->dispatcher->dispatchTyped($event);
-		} catch (\Throwable $e) {
-		}
+		$this->service->installBot($backend);
 	}
 }