+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 1.0.1 - 2023-12-13
+### Added
+- Added an OCC command to repair the bot
+
+### Fixed
+- Fixed a bug in the OCC command to delete a secret
+
 ## 1.0.0 - 2023-10-30
 ### Added
 - Nextcloud 28 compatibility